packages/core/pyproject.toml
[project]
name = "pystructconf-core"
version = "0.0.1"
description = "Core abstractions and data models for pyStructConf."
readme = "../../README.md"
requires-python = ">=3.10,<3.14"
license = { file = "../../LICENSE" }
authors = [{ name = "chingyaw" }]
dependencies = ["pyyaml>=6.0"]

[build-system]
requires = ["hatchling>=1.25"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/pystructconf_core"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"
